Add files via upload
At here, the parallel reaction model or distribution reaction model is used to deconvolute the DTG peaks. There are other models as well such as the Kissinger model, Coats and Redfern model, and isoconversional model. A comparison on the current dataset shows that they provide almost similar values of parameters. However, isoconversional model is assumption-free. That is why it is prefered some times. On the other hand, the parallel reaction model provides a sense of how the reaction proceeds.